Se pasa versión a ST1 para despliegue

+++ src/main/java/com/imss/sivimss/planfunerario/model/response/DatosConvenioResponse.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class DatosConvenioResponse {
 	
 	private String estado;
-	private Integer costoRenovacion;
+	private Double costoRenovacion;
 	private Integer indRenovacion;
 	private String segundoApellido;
 	private Integer estatusConvenio;
